Extract wait function into separate file
